chore: migrate API endpoints to improve structure and consistency (#48)
* chore: migrate API endpoints to improve structure and consistency

- Moved `v1/users/[userId]/zonesByCompany` to
  `v1/companies/[companyId]/zones` for better alignment with company-related data.
- Moved `v1/users/[userId]/userPlants` to `v1/companies/[companyId]` to centralize company-related endpoints.
- Removed outdated endpoints under `users/[userId]/` for companies, users, and zones.
- Updated and modified endpoint implementations for:
  - `v1/companies`
  - `v1/plants`
  - `v1/users`
  - `v1/zones`
  - `v1/users/[userId]/plants`

* chore: update services

* chore: update services used in components
