+++ src/app/dashboard/plants/page.tsx
@@ -1,12 +1,18 @@
 import { Suspense } from "react"
 import { auth } from "@/lib/auth"
-import { getPlantByCompany } from "@/lib/services"
+import { getPlantsByCompanyId, getUserById } from "@/lib/services"
 import { Table } from "@/ui/dashboard/plants/table"
 
-const DashboardPlantsPage = async () => {
+const getInformation = async () => {
     const session = await auth()
     const userId = session?.user?.id ? Number(session.user.id) : Number.MAX_SAFE_INTEGER
-    const plants = await getPlantByCompany(userId)
+    const { companyId } = await getUserById(userId)
+    const plants = await getPlantsByCompanyId(companyId)
+    return { plants }
+}
+
+const DashboardPlantsPage = async () => {
+    const { plants } = await getInformation()
 
     return (
         <section className="min-h-main py-4 space-y-4">
