+++ src/app/dashboard/samples/page.tsx
@@ -3,7 +3,7 @@ import { Suspense } from "react"
 import { auth } from "@/lib/auth"
 import { SampleList } from "@/ui/dashboard/samples/sample-list"
 import { Filter } from "@/ui/common/filter"
-import { getSamplesByUser, getZonesByUser } from "@/lib/services"
+import { getSamplesByUser, getUserById, getZonesByCompanyId } from "@/lib/services"
 
 export const metadata: Metadata = {
     title: "List of samples",
@@ -13,7 +13,8 @@ export const metadata: Metadata = {
 const getInformation = async () => {
     const session = await auth()
     const userId = session?.user?.id ? Number(session.user.id) : Number.MAX_SAFE_INTEGER
-    const [zones, samples] = await Promise.all([getZonesByUser(userId), getSamplesByUser(userId)])
+    const { companyId } = await getUserById(userId)
+    const [zones, samples] = await Promise.all([getZonesByCompanyId(companyId), getSamplesByUser(userId)])
     return { zones, samples }
 }
