+++ src/app/dashboard/users/page.tsx
@@ -1,12 +1,18 @@
 import { Suspense } from "react"
 import { auth } from "@/lib/auth"
-import { getUserByCompany } from "@/lib/services"
+import { getUserById, getUsersByCompanyId } from "@/lib/services"
 import { Table } from "@/ui/dashboard/users/table"
 
-const DashboardCompaniesPage = async () => {
+const getInformation = async () => {
     const session = await auth()
     const userId = session?.user?.id ? Number(session.user.id) : Number.MAX_SAFE_INTEGER
-    const users = await getUserByCompany(userId)
+    const { companyId } = await getUserById(userId)
+    const users = await getUsersByCompanyId(companyId)
+    return { users }
+}
+
+const DashboardCompaniesPage = async () => {
+    const { users } = await getInformation()
 
     return (
         <section className="min-h-main py-4 space-y-4">
