+++ src/lib/services/plants.ts
@@ -1,40 +1,13 @@
-import { Plants, UserPlants, Zones } from "@prisma/client"
+import { UserPlants } from "@prisma/client"
 import { getFetch } from "@/lib/utils"
 
 /**
- * Fetches all plants from the database by user
+ * Fetches the user plants related to the given user id.
  *
- * @param userId - The userId to fetch plants
- * @returns {Promise<Plants[]>} - A list of plants
+ * @param {number} userId - The user id.
+ * @returns {Promise<UserPlants[]>} - A promise that resolves to an array of UserPlants.
  */
-export const getPlantsByUser = async (userId: number): Promise<Plants[]> => {
-    const { data } = await getFetch<Plants[]>(`users/${userId}/plants`)
-    return data
-}
-
-/**
- * Fetches all zones from the database by the company associated with the user
- *
- * @param userId - the userId to fetch zones
- * @returns {Promise<Zones[]>} - A list of zones
- */
-export const getZonesPlantsByUser = async (userId: number): Promise<Zones[]> => {
-    const { data } = await getFetch<Zones[]>(`users/${userId}/zonesByCompany`)
-    return data
-}
-
-/**
- * Fetches all plants from the database by company
- *
- * @param userId - The userId to fetch plants
- * @returns {Promise<Plants[]>} - A list of plants
- */
-export const getPlantByCompany = async (userId: number): Promise<Plants[]> => {
-    const { data } = await getFetch<Plants[]>(`users/${userId}/plants`)
-    return data
-}
-
 export const getUserPlantByCompany = async (userId: number): Promise<UserPlants[]> => {
-    const { data } = await getFetch<UserPlants[]>(`users/${userId}/userPlants`)
+    const { data } = await getFetch<UserPlants[]>(`companies/${userId}`)
     return data
 }
