+++ src/lib/services/users.ts
@@ -1,17 +1,6 @@
-import { Users } from "@prisma/client"
+import { Samples, Users } from "@prisma/client"
 import { getFetch } from "@/lib/utils"
-
-/**
- * TODO: migrate the `getUserByCompany` function to the `company` module.
- *
- * Fetches all users associated with a specific company from the database
- * @param {number} userId - The ID of the user whose company users are to be fetched
- * @returns {Promise<Users[]>} - A promise that resolves to a list of users
- */
-export const getUserByCompany = async (userId: number): Promise<Users[]> => {
-    const { data } = await getFetch<Users[]>(`users/${userId}/users`)
-    return data
-}
+import { UsersResponse } from "../@types/types"
 
 /**
  * Fetches all users from the database
@@ -19,6 +8,28 @@ export const getUserByCompany = async (userId: number): Promise<Users[]> => {
  * @returns {Promise<Users[]>} - A list of users
  */
 export const getUsers = async (): Promise<Users[]> => {
-    const { data } = await getFetch<Users[]>(`users`)
+    const { data } = await getFetch<Users[]>("users")
+    return data
+}
+
+/**
+ * Fetches a specific user from the database by their id.
+ *
+ * @param {number} userId - The user id to get the user from the database
+ * @returns {Promise<UsersResponse>} - A user object
+ */
+export const getUserById = async (userId: number): Promise<UsersResponse> => {
+    const { data } = await getFetch<UsersResponse>(`users/${userId}`)
+    return data
+}
+
+/**
+ * Fetches the sample data associated with a specific user from the database.
+ *
+ * @param {number} userId - The user id to get the samples related to them from the database
+ * @returns {Promise<Samples[]>} - A list of samples related to the user
+ */
+export const getSamplesByUser = async (userId: number): Promise<Samples[]> => {
+    const { data } = await getFetch<Samples[]>(`users/${userId}/samples`)
     return data
 }
